+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.1.0] - 2018-12-14
+## [2.3.0] - 2019-01-16
+### Added
+- Matrix endpoint, docs, and tests
+- Map Matching endpoint, docs, and tests
+- Optimization endpoint, docs, and tests
+
+### Changed
+
+### Removed
+
+## [2.2.0] - 2018-12-14
 ### Added
 - Selection of directions profiles
 - Isochrone endpoint